+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Tests for "Conway sequence".
+- Tests for "Dwarfs standing on the shoulders of giants".
 
 ## [1.1.0] - 2022-02-28
 ### Added